refactor(abstract-eth): move txbuilder to abstract-eth WIN-1021
BREAKING CHANGE: getTransactionBuilder method is removed from EthLikeToken
class in abstract-eth module because TransactionBuilder in the
abstract-eth module is abstract class and hence cannot be instantiated. Hence the implementation of TransactionBuilder can
be added to the class that will inherit EthLikeToken class

TransactionPrebuild from new class AbstractEthLikeNewCoins is being
exported now instead of TransactionPrebuild from AbstractEthLikeCoin
class as the TransactionPrebuild from AbstractEthLikeNewCoins also has
support for hop transactions, batch transactions, etc

TICKET: WIN-1021
