+++ modules/abstract-eth/package.json
@@ -37,10 +37,16 @@
     ]
   },
   "dependencies": {
-    "@bitgo/sdk-coin-eth": "^4.10.0",
     "@bitgo/sdk-core": "^8.26.0",
     "@bitgo/statics": "^29.0.0",
     "@bitgo/utxo-lib": "^9.16.0",
-    "bignumber.js": "^8.0.1"
+    "bignumber.js": "^9.1.1",
+    "ethers": "^5.1.3",
+    "ethereumjs-util": "7.1.5",
+    "@ethereumjs/common": "^2.6.5",
+    "@ethereumjs/tx": "^3.3.0",
+    "ethereumjs-abi": "^0.6.5",
+    "bn.js": "^5.2.1",
+    "debug": "^3.1.0"
   }
 }
