+++ modules/abstract-eth/src/index.ts
@@ -2,3 +2,5 @@ export * from './abstractEthLikeCoin';
 export * from './ethLikeToken';
 export * from './abstractEthLikeMPCCoin';
 export * from './ethLikeMPCToken';
+export * from './lib';
+export * from './abstractEthLikeNewCoins';
