+++ modules/abstract-eth/src/lib/index.ts
@@ -1,0 +1,17 @@
+export * from './contractCall';
+export * from './iface';
+export * from './keyPair';
+export * from './transaction';
+export * from './transactionBuilder';
+export * from './transferBuilder';
+export * from './transferBuilders';
+export * from './transferBuilder';
+export * from './types';
+export * from './utils';
+export * from './walletUtil';
+
+// for Backwards Compatibility
+import * as Interface from './iface';
+import * as Utils from './utils';
+
+export { Interface, Utils };
