+++ modules/abstract-eth/src/lib/transferBuilders/transferBuilderERC1155.ts
@@ -1,4 +1,6 @@
 import { BuildTransactionError, InvalidParameterValueError } from '@bitgo/sdk-core';
+import { hexlify, hexZeroPad } from 'ethers/lib/utils';
+
 import { ContractCall } from '../contractCall';
 import { decodeERC1155TransferData, isValidEthAddress, sendMultiSigData } from '../utils';
 import {
@@ -8,7 +10,6 @@ import {
   ERC1155SafeTransferTypeMethodId,
 } from '../walletUtil';
 import { BaseNFTTransferBuilder } from './baseNFTTransferBuilder';
-import { hexlify, hexZeroPad } from 'ethers/lib/utils';
 
 export class ERC1155TransferBuilder extends BaseNFTTransferBuilder {
   private _tokenIds: string[];
