+++ modules/abstract-eth/tsconfig.json
@@ -4,15 +4,12 @@
     "outDir": "./dist",
     "rootDir": "./",
     "strictPropertyInitialization": false,
-    "esModuleInterop": false,
+    "esModuleInterop": true,
     "typeRoots": ["../../types", "./node_modules/@types", "../../node_modules/@types"]
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules"],
   "references": [
-    {
-      "path": "../sdk-coin-eth"
-    },
     {
       "path": "../sdk-core"
     },
