+++ modules/sdk-coin-eth/src/lib/transferBuilders/index.ts
@@ -1,3 +1,3 @@
-export * from './baseNFTTransferBuilder';
-export * from './transferBuilderERC1155';
-export * from './transferBuilderERC721';
+import { BaseNFTTransferBuilder, ERC721TransferBuilder, ERC1155TransferBuilder } from '@bitgo/abstract-eth';
+
+export { BaseNFTTransferBuilder, ERC721TransferBuilder, ERC1155TransferBuilder };
