+++ modules/sdk-coin-eth/tsconfig.json
@@ -11,6 +11,9 @@
   "exclude": ["node_modules"],
   "references": [
     {
+      "path": "../abstract-eth"
+    },
+    {
       "path": "../sdk-core"
     },
     {
